CLOUDERA-BUILD. Preview of SPARK-3606.
[SPARK-3606] [yarn] Correctly configure AmIpFilter for Yarn HA (1.1 version).

This is a backport of SPARK-3606 to branch-1.1. Some of the code had to be
duplicated since branch-1.1 doesn't have the cleanup work that was done to
the Yarn codebase.

I don't know whether the version issue in yarn/alpha/pom.xml was intentional,
but I couldn't compile the code without fixing it.

(cherry picked from commit db67825813aab8236cdc435ef3bae1fa131a3119)