CLOUDERA-BUILD. Exclude more jars from the assembly.
The assembly jar is big; so big that it's currently over 65k
files, the limit where some tools start breaking when trying
to read the file. Even some JDK tools seem to not like that.

So add a few more exclusions when "-Phadoop-provided" is defined.
These are parquet and hive libraries already shipped with other
CDH packages. The only end-user effect is that someone using
Spark SQL + Hive will have to manually add Hive jars to the
classpath, since those are not included in the output of
"hadoop classpath".

(cherry picked from commit 2037c2316aeb839280e79fe7aff76d8002fddd20)