CDH-33655. Use version of Spark defined in user configuration.
During rolling upgrades, the activated parcel and the active client
configuration in /etc/spark may not match. This may cause jobs to fail to
launch because the config is tailored for a particular Spark version.

This change detects when the versions do not match and runs Spark from the
version defined in the configuration, warning the user that this is
happening.

(cherry picked from commit 08476d357043971ed893c6d5677e76de3d300f19)