[SPARK-12696] Backport Dataset Bug fixes to 1.6
We've fixed a lot of bugs in master, and since this is experimental in 1.6 we should consider back porting the fixes.  The only thing that is obviously risky to me is 0e07ed3, we might try to remove that.

Author: Wenchen Fan <wenchen@databricks.com>
Author: gatorsmile <gatorsmile@gmail.com>
Author: Liang-Chi Hsieh <viirya@gmail.com>
Author: Cheng Lian <lian@databricks.com>
Author: Nong Li <nong@databricks.com>

Closes #10650 from marmbrus/dataset-backports.

(cherry picked from commit a6190508b20673952303eff32b3a559f0a264d03)