[SPARK-13642][YARN][1.6-BACKPORT] Properly handle signal kill in ApplicationMaster
## What changes were proposed in this pull request?

This patch is fixing the race condition in ApplicationMaster when receiving a signal. In the current implementation, if signal is received and with no any exception, this application will be finished with successful state in Yarn, and there's no another attempt. Actually the application is killed by signal in the runtime, so another attempt is expected.

This patch adds a signal handler to handle the signal things, if signal is received, marking this application finished with failure, rather than success.

## How was this patch tested?

This patch is tested with following situations:

Application is finished normally.
Application is finished by calling System.exit(n).
Application is killed by yarn command.
ApplicationMaster is killed by "SIGTERM" send by kill pid command.
ApplicationMaster is killed by NM with "SIGTERM" in case of NM failure.

Author: jerryshao <sshao@hortonworks.com>

Closes #11690 from jerryshao/SPARK-13642-1.6-backport.

(cherry picked from commit 5e9cefc8ccfaa0ef0bb0f2052f9aa755197b0184)