CLOUDERA-BUILD. Avoid downloading dependencies in Hive metastore tests.
Use the "builtin" jars for tests, since we want to avoid ivy downloading
hadoop dependencies during the build; also because CDH's hadoop libraries
are not in maven central, and so we get ugly log messages in the test
output.

I also disabled a test that becomes redundant with these changes, and another
that relies on configuration that causes the download code to trigger (and
was failing, maybe because of that, except the exception was a little odd).

(cherry picked from commit 10d80e54acf6cca90b28e4107db23e4d4edd233c)