CDH-42136. Enforce version match between Spark and CSD.
To avoid an exploding test matrix, let's enforce that the
CSD being used matches the version of the parcel being used.
This is done by writing the CSD version to a file that is
checked by the control script.

The message formatting is slightly odd because when CM shows
logs it does not use a monospaced font, so ASCII art does not
show properly.

As part of this change, I had to modify all version so use a
period instead of a dash to separate "cloudera1", so that it's
easy to have the version string be the same across the board
(packaging and maven).