[SPARK-16953] Make requestTotalExecutors public Developer API to be consistent with requestExecutors/killExecutors
## What changes were proposed in this pull request?

RequestExecutors and killExecutor are public developer APIs for managing the number of executors allocated to the SparkContext. For consistency, requestTotalExecutors should also be a public Developer API, as it provides similar functionality. In fact, using requestTotalExecutors is more convenient that requestExecutors as the former is idempotent and the latter is not.

Author: Tathagata Das <tathagata.das1565@gmail.com>

Closes #14541 from tdas/SPARK-16953.

(cherry picked from commit 8650239050ade91689ffa0672ea094de2594e37c)
Signed-off-by: Tathagata Das <tathagata.das1565@gmail.com>
(cherry picked from commit 9748a29289cf35a739854f17e72d039009abedbe)